requirements.txt
litellm>=1.0.0
python-dotenv>=1.0.0
jinja2>=3.0.0
pyyaml>=6.0.0
flask>=3.0.0

